Add option to set a cookie name
With an external server managing authentication, we need to match
whatever the server is setting on the headers.

Then, we don't need to use a `verify` function on the `options`, I think
an empty function would work for now.

Some tests got deleted since I don't think we need to test other
libraries functionalities, at least what they are intended to do.