Add option to know if we are using external server
When using the `verify` function to authenticate a user inside our
Next.js app we encrypt the user object and can decrypt it when we need
to, but when we use an external server and that server is setting the
cookie header itself, we don't have control over how the cookie other
than read it and check if it's present.

That's why we check if we are using an external server and if we do,
then we don't attempt to decrypt a user object since it won't exist.